requirements.txt
# Core Deep Learning (CUDA 11.3 compatible)
# Note: It is recommended to install PyTorch manually first if CUDA issues arise.
torch==1.12.1
torchvision==0.13.1
pytorch-lightning==1.9.5
torchmetrics==0.11.4

# Numerical & Data Processing
numpy==1.21.6
scipy==1.7.3
pandas==1.3.5
scikit-learn==1.0.2

# Computer Vision & Image Augmentation
opencv-python==4.6.0.66
Pillow==9.5.0
scikit-image==0.19.3
imgaug==0.4.0

# 3D Processing & Geometry
open3d==0.17.0
shapely==2.0.7

# Simulation & Traffic (V2X related)
traci==1.22.0
sumolib==1.22.0

# Visualization
matplotlib==3.5.3

# Utilities
tqdm==4.67.1
PyYAML==6.0.1
addict==2.4.0
fire==0.7.0